+++ packages/api/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @atproto/api
 
+## 0.16.4
+
+### Patch Changes
+
+- Updated dependencies [[`2104d9033`](https://github.com/bluesky-social/atproto/commit/2104d9033e2e1a3a7b821c1f0c5c8ffac5832d59)]:
+  - @atproto/lexicon@0.4.14
+  - @atproto/xrpc@0.7.3
+
 ## 0.16.3
 
 ### Patch Changes