+++ packages/lexicon-resolver/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @atproto/lexicon-resolver
 
+## 0.1.1
+
+### Patch Changes
+
+- Updated dependencies [[`2104d9033`](https://github.com/bluesky-social/atproto/commit/2104d9033e2e1a3a7b821c1f0c5c8ffac5832d59)]:
+  - @atproto/lexicon@0.4.14
+  - @atproto/repo@0.8.7
+  - @atproto/xrpc@0.7.3
+
 ## 0.1.0
 
 ### Minor Changes